Scrap my unused code, big commit = big description
Sorry for this big commit, I'm reusing some code that I used for another project.

Updated go.mod to change the path of the module and to add dependencies.

The main function is getting the port to listen to from getEnv() giving a path to the .env file. It then gets the router that will server as the http handler from getRouter(). It then create a http server and will start to listen for connections. If it can't listen, the error will be printed.

The get_env.go file : the getEnv() function gets a path to an env file as an option and will the source it, print an error if it can't. It will then read the port and the postgres database URL and will check them with envCheck() it prints an error if they are invalid. otherwise they will be returned. The envCheck function will get a port and a postgres database URL and check them. For the port, it will convert it to a number and check if it is null or invalid and return an error if it is the case. For the database, I'm only checking if it is null, if it is, return an error.

The router.go file : the createRouter() function will create a chi router and return it. The configRouter() function will configure the router to accept http and https + the GET and POST methods, it also accepts all headers but no http creds. The request is cached for 300sec. The subRouters() function will create and return sub-routers that will be mounted on the main router from createRouter() currently there is only one sub-router to handle the '/ok', '/err' and '/' endpoints. The '/ok' is basically a health check, '/err', is another health check just to check if errors works and '/' is a static redirection to "https://go.dev/", it will then return said sub-router. The mountRouter() function will mount given sub-routers the main router at the '/' mount point and return it. The getRouter() function is basically the driver for all of those functions and will return the main router().

The json.go file : the respondWithError() function will serve an error message in json to the client based on an error code and an error message. The respondWithJSON() will server a JSON payload based on a code and an interface, it will Marshal (convert in json) the payload and serve it to the user. If there is an error in the process, it will be returned as a server-side error.

The handler_err.go file : the handleErr() function will call respondWithError() giving it a client-side error code and a basic error message. It is being called when the client requests the '/err' endpoint.

The handler_readiness.go file : the handlerReadiness() function will call respondWithJSON() giving at a success code and an empty message. It is being called when the client requests the '/ok' endpoint.

The handler_redirect.go file : the handlerRedirect() function redirects the client to another static URL, currently "https://go.dev". It is being called when the client requests the '/' endpoint.

Also added the vendor directory in case of dependency errors, there would be a local version of the deps.