Add handler_link.go witch handles the creation of link entries and calling handlerRedirect() specifying a link
The handlerCreateLink() function will decode a JSON payload sent by the client and extract information from it :
- A url to point to <- required
- The length for a random short (max 16, default of 6) <- optional
- A custom short (max length of 16) <- optional
- An expiry time in minutes (-1 = never, default of 48 hours)
It will then check and create defaults for this information and will create a database entry based on them.

The handlerGetLink() function will call handlerRedirect() specifying it a url associated with a short (given by getURL()) that will in turn redirect the client to the url.