Use self-made sql instead of auto generated sql with sqlc and goose
Although there is no wrong in using them, it is a bit overkill for this project.

Essentially, the new functions in links.go are basically simplified clone of the sqlc generated ones, but less complicated with two more functions:
- checkTable(): check if a given table exists with the database.
- CreateLinksTable(): create a the 'links' table.
The "clones" are:
- CreateLink(): Create an entry in the links table comprised of a given uuid, creation time, expiration time, an url, and a short.
- GetLinks(): Get a list of strings entries comprised of only the expiration time, the url and the short, other information are useless.
- RemoveLink(): Remove of link entry from the links table based on its short.

db.go replaces db_api.go for connecting the database in a simpler manner.

models.go (in database) replaces models.go (in main), only the Link struct remains.

router.go and the handlers are modified to reflect those changes, they are essentially the same.

handlerRedirectToUrl() (previously handlerGetLink()) now replaces handlerRedirect(), it gets the short from the url path and redirect to the corresponding url.

Removed get_url.go, too complicated for such a basic thing a getting the url path, doing it handlerRedirectToUrl() instead.

Created a dedicated util.go for trimFirstRune().

Fixed 404 typo.