Simplify the handlers by not using chi routers anymore
Handlers are now only managed by http.HandleFunc, the createLink() and redirectToUrl() are now simple functions, called by the handlerRoot() handler which decides whether to use one or the other based solely on the method, GET will call redirectToUrl() and POST will call createLink().

The other handlers now check if the method is valid, calling respondWithError() with a 405 if the method isn't valid.

Errors now show the error code in the response, previously it was only given in the header.

The database connection and the handler to path assignment are now done in main.go.

More comments, optimized ListenAndServe.