Add passwords
CreateLinksTable() adds a password column for an optional argon2 hash.

CreateLink() adds said optional hash along with the other data.

The new function GetHashByShort() gets the argon2 hash associated with a given short.

createLink() changes: If a password is given during a POST at "/", we will make an argon2 hash and store it in the database.

redirectToUrl() changes: When a GET request to a short is made, it will check if the short has a hash associated with it. It will then compare the hash against a given password in json. If the password is wrong or not password is given, there will be an error.

Minor tweak in createLink(): the month of December has 31 days (max expiration time).