Add a user interface
.env.example: Added INSTANCE_NAME and INSTANCE_URL to the .env variables, they will be used in the user interface. There is also an example of how the postgres URL should look like and also, the time between the database cleanups is configurable with TIME_BETWEEN_DB_CLEANUPS (in minutes). NOTE: The database configuration will drastically change in a near future update.

get_env.go: Changed the functions to reflect the new .env changes:
- getEnv() now gets INSTANCE_NAME, INSTANCE_URL and TIME_BETWEEN_DB_CLEANUPS from .env.
- envCheck() now checks whether the INSTANCE_NAME is null, if it is, throw an error, also checks the URL, if the INSTANCE_URL is null or does not correspond to a valid URL, throw an error. The time between cleanups is also checked, if it is null, throw an error. The DB_URL also checks for a valid postgres URL. NOTE: The database configuration checks will drastically change in a near future update.

main.go: Get all these new configurations from getEnv() and use a new struct sendToHandlers{} (declared in utils.go) to hold the database connection, the instance name and URL which will be stored in handlersInfo. Also, though it is rare and not really necessary to do so, the closing of the database connection is deferred. handlersInfo will then be passed to the handlers that needs them. The garbage collection now gets the time between cleanups as an argument.

util.go was renamed to utils.go.

utils.go: The new sendToHandlers{} is declared here, and the parameters struct from handler_link.go has been moved here. collectGarbage() uses the new struct and the timeBetweenCleanups arg that will set the duration of the sleep in minutes. There is also a new function, decodeJSON() which is used to decode JSON parameters from the body of the request. It basically does what createLink() did at the start but in a separate function.

handler_link.go was renamed to handler_api.go to make a clear distinction between front and api.

handler_api.go: The order of function where change to a more logical one i.e. declare before calling, not obligatory, but I prefer it like that. redirectToUrl() is now renamed to apiRedirectToUrl().
- apiRedirectToUrl() now checks for a JSON header to decide whether to throw an error for the lack of password or serve the user for a UI to get the password by calling frontAskForPassword() (declared in handler_front.go). It also checks if the password is given as a parameter in the URL and use it as such. Though not the recommended way to do it, it works. createLink() was renamed to apiCreateLink().
- apiCreateLink() now uses parameters given by decodeJSON() (declared in utils.go) to check and use them to create a record in the database. It returns the response from the database, an error code, 0 if everything is fine, and an empty string if everything is fine. If there is an error, it will return the error code and the user apropriate error message for it.
- apiHandlerRoot(): calls correct function depending on the method or request. if there is a request for /favicon.ico, drop it to avoid unnecessary 404 errors. If the client requests "/" with GET, call frontHandlerMainPage() (declared in handler_front.go). If the user requests any other path with GET, call apiRedirectToUrl(). If the client requests "/" with POST, get the JSON parameters with decodeJSON() and send it to apiCreateLink(). respond with an error in case of errors, send back the response of the database query if everything is fine. Finally, if no requests meets the conditions, it's not an allowed method.

handler_front.go: Define a Page{} struct for the content of the pages. Declare a global 'templates' variables to parse the HTML templates in advance. This file adds new functions:
- renderTemplate(): Execute a template from already parsed ones using a name and the content (Page{}) given as arguments. JSON error in case of errors.
- frontErrorPage(): Serve an error page based on given arguments.
- frontHandlerMainPage(): Serve a front page with a form for shortening a link.
- frontCreateLink(): Near-exact clone of apiCreateLink, only the errors changes by calling frontErrorPage().
- frontHandlerAdd(): Calls frontCreateLink() giving it the values of the form of the main page, then serve the response from the database to the user in a readable format.
- frontAskForPassword(): Serve the user with a page that asks for a password.
- frontHandlerRedirectToUrl: Near-exact clone of apiRedirectToUrl(), it gets the password and the short, then compares the password and the hash to then approve and redirect to the corresponding url.

Added the HTML and assets for the UI.

Removed the creation of a front/UI from the TODO list and add the support of more SQL db engines and potential error fix to the list.