Add more error-proof auto-generation
When auto-generating a path, if the path is already taken, instead of blaming the user, we regenerate new ones until it's correct, we inform the user in case the length he specified was changed in the process. Not totally error proof but it lessens the probability of an unexpected error by a lot.