Introduce the use of tokens for very tiny and vague security reasons
A new global token is generated every 30 seconds.

The token is used in the forms of index.html and pass.html (respectively "/" and "/access" endpoints), the time of request is also hidden.

When the token is invalid and the time of the request was more than 15 minutes ago, throw an error.