Add unit testing
Gradually moving to TDD, test are done with testify, the tests do not concern the main handlers for the time being.

Although tests are not written for the main handlers (this is temporary), it is generally safe to assume that everything works since the underlying functions used by the handlers are tested, however any changes made to the main handlers does not guaranty this stays the case.