Stop using testify
Since only a small fraction of testify features are used in the tests, I elected to stop using it.

Added the "assert" helper I've made for git.xdol.org/xdol/hashdog and modified it for feature parity with the features that were used with testify.

There is no logic difference between tests done with testify and tests done using assert.

Also reorganized the tests.