Embed static files
The reddlinks binary is now portable, (you either need a .env file or you can set env variables yourself).

Using 'embed', the static/ dir is embedded in the binary

Instead of parsing each html files on the disk, they are parsed from the embed fs

The assets have their own filesystem, which is served the same way as before

Reflect this change in the Dockerfile.