Add user-friendly expiry time input
Use gitlab.gnous.eu/ada/atp to parse time from json input instead of using a number of minutes which can be confusing.

Use a datetime-local input for the front, which is parsed and used as the expiry date. For some reason, gostd and html datetime-local uses different rfc3339 formats.

Small tweaks on index.html and main.css