Use modernc.org/sqlite to build without CGO
Replaced github.com/mattn/go-sqlite3 with modernc.org/sqlite, to be able to build reddlinks without cgo when using an sqlite database. Only downside is that this driver has a lot of dependencies.

Use go build instead of make in the Dockerfile, by not having to install it, the building time of the image is a bit faster than before.

Small tweaks.