Add info shortened link info page
If the suffix '+' is used at then end of a shortened path, an information page will be displayed. Password protection is not bypassed.

For CLI clients, a JSON response is given, for browsers, an HTML response is given.

Added a function to get url, creation date and expiration date of a given short from the database.