Redo static and locales system
Use only one set of templates.

Add GetLocales function in utils which returns a map of locales based on the json files present in the static dir. It also return a slice of supported languages.

The locales files are JSON files containing keys and translations for all of the text present in the pages except for errors.

The corresponding locale to display is chosen in RenderTemplate based on the Accept-Language header.

Page struct is renamed to PageParameters.

Fixed golangci conf.

Formatting.